+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# vTBD
+- [BP-1545](https://movai.atlassian.net/browse/BP-1545): Remove old Alerts implementation
+
 # v3.14.0
 - [BP-1460](https://movai.atlassian.net/browse/BP-1460): API to expose fleet information